moved the get number of annotations function to a new analysis file
