+++ pigeonXT/__init__.py
@@ -1,1 +1,2 @@
 from .annotate import *
+from .analysis import *
