+++ pigeonXT/annotate.py
@@ -206,12 +206,3 @@ def annotate(examples,
     next_example()
     return annotations
 
-
-def get_number_of_annotations_per_label(annotations, labels):
-    annotations_count = {}
-    chosen_labels_list = [annotated_labels for annotated_labels in annotations.values()]
-
-    for label in labels:
-        annotations_count[label] = sum(label in chosen_labels for chosen_labels in chosen_labels_list)
-
-    return annotations_count
